Move validation_config to data
