+++ src/cupbearer/scripts/conf/train_classifier_conf.py
@@ -1,42 +1,13 @@
-import dataclasses
 import os
 from dataclasses import dataclass
 from typing import Optional
 
-from cupbearer.data import DatasetConfig, NoData
+from cupbearer.data import DatasetConfig, ValidationConfig
 from cupbearer.models import CNN, MLP, ModelConfig
-from cupbearer.utils.config_groups import (
-    config_group,
-    register_config_group,
-)
+from cupbearer.utils.config_groups import config_group
 from cupbearer.utils.optimizers import Adam, OptimizerConfig
 from cupbearer.utils.scripts import DirConfig, ScriptConfig
-from cupbearer.utils.utils import BaseConfig
 from simple_parsing.helpers import mutable_field
-
-
-@dataclass(kw_only=True)
-class ValidationConfig(BaseConfig):
-    val: Optional[DatasetConfig] = config_group(DatasetConfig, NoData)
-    clean: Optional[DatasetConfig] = config_group(DatasetConfig, NoData)
-    backdoor: Optional[DatasetConfig] = config_group(DatasetConfig, NoData)
-
-    def items(self) -> list[tuple[str, DatasetConfig]]:
-        res = []
-        for field in dataclasses.fields(self):
-            value = getattr(self, field.name)
-            if isinstance(value, DatasetConfig) and not isinstance(value, NoData):
-                res.append((field.name, value))
-
-        return res
-
-
-register_config_group(
-    ValidationConfig,
-    {
-        "default": ValidationConfig,
-    },
-)
 
 
 @dataclass(kw_only=True)
